requirements.txt
langgraph==0.2.50
ipykernel
langchain_community==0.3.31
langchain_core>=0.3.68,<0.4.0
tavily-python==0.7.17
wikipedia==1.4.0
langchain_openai==0.3.32
langchain-google-genai==2.1.8
langchain-groq==0.3.6
structlog==25.4.0
python-docx==1.2.0
reportlab==4.4.7
-e .
